Update the score based on the answer
Updated the code to checked if the user selected answer is equal to the
correct answer of the current question. If it is true add one to the
current count of correct answers and update the score. If false, just
update the score based on count of correct answers and count of
questions answered.