Can't display method generate random answer
Issue: An error was made when a method was use in the rendor() to
randomly select an answer and update the Redux state.

The soluction was to create a new state to hold the random answer. The
random answer is generated and save to the new state when the component
loads and when the user press the DoneButton component.  The random
answer is saved to userAnswer state to disable and enable the
CheckAnswersButton.