Bug: changing answers on MCAnswersList
Issues: On the MCAnswersList page the answers would change. Fix changing
answers after pushing the checkAnswerButton.
Solution: Use Redux to save the three answers to the store and then push
them to be displayed. The CSS had to be tweek on the local level instead
of depending on changing the entire array of questions/answers using
redux.